feat(api): :sparkles: add register user use-case interactor
The interactor is an use-case. It receives an request model obj from the controller, it runs the logic e.g. validation. After that it will do database operation through an user repository gateway and send an response model through the output boundary to the presenter.
The provided registerGuest method will:

 - Validate the input data.
 - Check if username and email are still available.
 - Create the user in the db through the repository interface.
 - Send an email with a generated verification token through the notification service.
 - And finally send the response to the presenter.